Cleanup and refactors
